Revert "Revert "refactor of utils and versioning code""
This reverts commit 3b8abd4aae0ed65df56ba99e05819755da56e2b9.
