+++ app/docs/[...slug]/DocumentPageClient.tsx
@@ -4,10 +4,10 @@ import { useTina } from "tinacms/dist/react";
 import { useScreenResizer } from "../../../components/hooks/ScreenResizer";
 import { TinaMarkdown } from "tinacms/dist/rich-text";
 import { DocsMDXComponentRenderer } from "../../../components/tinaMarkdown/DocsMDXComponentRenderer";
-import { formatDate } from "../../../utils/docs/getFormattedDate";
+import { formatDate } from "../../../utils/generic/getFormattedDate";
 import DocsPagination from "../../../components/ui/Pagination";
 import MainDocsBodyHeader from "../../../components/docs/MainDocsBodyHeader";
-import { useTocListener } from "../../../utils/docs/tocListener";
+import { useTocListener } from "../../../utils/navigation/tocListener";
 import ToC from "../../../components/docs/PageToc";
 import { LeftHandSideParentContainer } from "../../../components/docs/LeftHandSideParent";
 import TocOverflowButton from "../../../components/docs/ToCOverflow";
