+++ components/docs/DocsNavigationList.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import styled, { css } from "styled-components";
 import { usePathname } from "next/navigation";
-import { matchActualTarget } from "../../utils/docs/urls";
+import { matchActualTarget } from "../../utils/navigation/urls";
 import { DynamicLink } from "../ui/DynamicLink";
 import { BiChevronRight } from "react-icons/bi";
 import AnimateHeight from "react-animate-height";
