+++ tina/collections/siteConfig.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-import { Collection } from "tinacms";
 import { CustomColorToggle } from "../../components/ui/CustomColorToggle";
 
 export const globalSiteConfiguration = {
