+++ utils/generic/sanitizeLabel.ts
