+++ utils/versioning/addSubpath.ts
@@ -1,0 +1,6 @@
+export const addSubpathToSlug = (slug: string, versionNumber: string) => {
+    return slug.replace(
+      /^(.+?\/docs\/)/,
+      `$1_versions/${versionNumber}/`
+    );
+  };
